containers/docker: allow async updates to complete after RPC ends
When the RPC ends the context will be cancelled, which will cause
m.performContainerUpdate to fail any calls which involve ctx.

To avoid this, the context cancellation is removed (overridden).
However to ensure the caller has still got some control over this
RPC's lifespan, any deadline on the context will be maintained.

Also log whether the update was successful or had an error