containers/docker: set valid maximum timeout for force-stopping container
If the ctx deadline is used as a timeout but the container
does not handle SIGTERM, then by the time SIGKILL is sent,
the ctx will have expired.
To fix this, spend at most half of the ctx deadline waiting
before issuing the SIGKILL.

To ensure that this does not take too long (as the Force
fields intention is to forcibly kill the container), the
timeout is capped at a maximum of 10 seconds (which is the
default timeout used by docker).

Also update the test to make it deterministic.
Previously the test could fail if the time.Now() in
ContainerStop called too late.